+++ tests/Unit/DependencyInjection/ConfigurationTest.php
@@ -395,4 +395,22 @@ class ConfigurationTest extends AbstractExtensionConfigurationTestCase
         $file = __DIR__.'/../../Resources/Fixtures/config/cache_config_with_no_pool.yml';
         $this->assertProcessedConfigurationEquals([], [$file]);
     }
+
+    public function testLimitlessCapturedBodyLength()
+    {
+        $file = __DIR__.'/../../Resources/Fixtures/config/limitless_captured_body_length.yml';
+        $config = $this->emptyConfig;
+        $config['profiling']['captured_body_length'] = null;
+        $this->assertProcessedConfigurationEquals($config, [$file]);
+    }
+
+    /**
+     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
+     * @expectedExceptionMessage The child node "captured_body_length" at path "httplug.profiling" must be an integer or null.
+     */
+    public function testInvalidCapturedBodyLengthString()
+    {
+        $file = __DIR__.'/../../Resources/Fixtures/config/invalid_captured_body_length.yml';
+        $this->assertProcessedConfigurationEquals([], [$file]);
+    }
 }